+++ conftest.py
@@ -75,7 +75,6 @@ def set_configs():
         let pyodide = await loadPyodide({
             fullStdLib: false,
             jsglobals : self,
-            enableRunUntilComplete: true,
         });
         """,
     )
@@ -90,7 +89,6 @@ def set_configs():
             fullStdLib: false,
             jsglobals: self,
             _loadSnapshot: snap,
-            enableRunUntilComplete: true,
         });
         """,
     )