+++ src/js/pyodide.ts
@@ -178,7 +178,8 @@ export async function loadPyodide(
      */
     pyproxyToStringRepr?: boolean;
     /**
-     * Make loop.run_until_complete() function correctly using stack switching
+     * Make loop.run_until_complete() function correctly using stack switching.
+     * Default: ``true``.
      */
     enableRunUntilComplete?: boolean;
     /**
@@ -221,7 +222,7 @@ export async function loadPyodide(
     env: {},
     packageCacheDir: indexURL,
     packages: [],
-    enableRunUntilComplete: false,
+    enableRunUntilComplete: true,
     checkAPIVersion: true,
     BUILD_ID,
   };
@@ -292,7 +293,7 @@ If you updated the Pyodide version, make sure you also updated the 'indexURL' pa
     snapshotConfig,
     options._snapshotDeserializer,
   );
-  API.sys.path.insert(0, API.config.env.HOME);
+  API.sys.path.insert(0, "");
 
   if (!pyodide.version.includes("dev")) {
     // Currently only used in Node to download packages the first time they are