+++ docs/project/changelog.md
@@ -23,6 +23,11 @@ _June 04, 2025_
 - {{ Fix }} Add the current working directory to the path instead of `$HOME`.
   {pr}`5630`
 
+- {{ Enhancement }} `pyodide.loadPackage` now prints the output to the `stdout`
+  and `stderr` streams that are passed to `loadPyodide()` or by
+  `pyodide.setStdout()` and `pyodide.setStderr()`.
+  {pr}`5621`
+
 ## Version 0.27.6
 
 _May 15, 2025_