+++ docs/project/changelog.md
@@ -22,6 +22,9 @@ _June 04, 2025_
   {pr}`5599`
 - {{ Fix }} Add the current working directory to the path instead of `$HOME`.
   {pr}`5630`
+- {{ Fix }} Fixed a fatal error when stack switching is enabled on a function
+  that raises an asynchronous error.
+  {pr}`5678`
 
 - {{ Enhancement }} `pyodide.loadPackage` now prints the output to the `stdout`
   and `stderr` streams that are passed to `loadPyodide()` or by